👩‍🏫

+++ board/src/main/java/com/joeun/board/controller/FileController.java
@@ -77,12 +77,19 @@ public class FileController {
 
 
     @DeleteMapping("")
+    // public ResponseEntity<String> deleteFile(@RequestBody Files file) throws Exception {
     public ResponseEntity<String> deleteFile(Files file) throws Exception {
+        log.info("[DELETE] - /file");
         int fileNo = file.getFileNo();
         log.info("fileNo : " + fileNo);
+        if( fileNo == 0 )
+            return new ResponseEntity<String>("FAIL", HttpStatus.BAD_REQUEST);       
 
         int result = fileService.delete(fileNo);
 
+        if( result == 0 )
+            return new ResponseEntity<String>("FAIL", HttpStatus.OK);    
+        
         return new ResponseEntity<String>("SUCCESS", HttpStatus.OK);
     }
